Move make_wrapper to things.py
